requirements.txt
Flask
PyYAML
pyperclip